+++ tools/replay/ReplayFile.hpp
@@ -10,6 +10,7 @@
 
 #include <string>
 #include <vector>
+#include "ReplayOptions.hpp"
 #include "umpire/Allocator.hpp"
 
 class ReplayFile {
@@ -35,6 +36,7 @@ public:
 
   struct AllocatorTableEntry {
     rtype type;
+    std::size_t line_number;    // Causal line number of input file
     bool introspection;
     char name[max_name_length];
     char base_name[max_name_length];
@@ -53,16 +55,6 @@ public:
         std::size_t min_alloc_size;
         int alignment;
       } pool;
-      struct {
-        std::size_t initial_alloc_size;
-        std::size_t min_alloc_size;
-        int alignment;
-      } dynamic_pool_list;
-      struct {
-        std::size_t initial_alloc_size;
-        std::size_t min_alloc_size;
-        int alignment;
-      } dynamic_pool_map;
       struct {
         std::size_t capacity;
       } monotonic_pool;
@@ -104,6 +96,7 @@ public:
 
   struct Operation {
     otype       op_type;
+    std::size_t op_line_number;     // Causal line number of input file
     int         op_allocator;
     void*       op_allocated_ptr;
     std::size_t op_size;            // Size of allocation/operation
@@ -121,7 +114,7 @@ public:
           | static_cast<uint64_t>('A') << 8
           | static_cast<uint64_t>('Y'));
 
-  const uint64_t REPLAY_VERSION = 11;
+  const uint64_t REPLAY_VERSION = 14;
 
   struct Header {
     struct Magic {
@@ -134,15 +127,17 @@ public:
     Operation ops[1];
   };
 
-  ReplayFile( std::string input_filename, std::string binary_filename );
+  ReplayFile( const ReplayOptions& options );
   ~ReplayFile( );
   ReplayFile::Header* getOperationsTable();
 
   void copyString(std::string source, char (&dest)[max_name_length]);
   bool compileNeeded() { return m_compile_needed; }
-  const std::string m_input_filename;
+  std::string getLine(std::size_t lineno);
+  std::string getInputFileName() { return m_options.input_file; }
 
 private:
+  ReplayOptions m_options;
   Header* m_op_tables{nullptr};
   const std::string m_binary_filename;
   int m_fd;